feat(desktop): qualify unsigned preview release
