+++ .github/actions/desktop-preview-candidate/action.yml
@@ -1,0 +1,134 @@
+name: Build and qualify desktop preview candidate
+description: Builds, proves, qualifies, and uploads one exact unsigned desktop preview artifact.
+
+inputs:
+  artifact-name:
+    description: Uploaded candidate artifact name.
+    required: true
+  source-sha:
+    description: Exact source revision recorded in release evidence.
+    required: true
+
+runs:
+  using: composite
+  steps:
+    - name: Set up Bun
+      uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2
+      with:
+        bun-version: 1.3.7
+
+    - name: Set up Go
+      uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16 # v6
+      with:
+        go-version-file: go.mod
+        cache: true
+
+    - name: Test native Windows policy probe
+      if: runner.os == 'Windows'
+      shell: pwsh
+      run: go test ./desktop/native/windowspolicy
+
+    - name: Install Debian package tooling
+      if: runner.os == 'Linux'
+      shell: bash
+      run: sudo apt-get update && sudo apt-get install --yes fakeroot
+
+    - name: Install pinned desktop dependencies
+      shell: bash
+      working-directory: desktop
+      run: bun install --frozen-lockfile
+
+    - name: Test preview and desktop security contracts
+      shell: bash
+      working-directory: desktop
+      run: bun run test
+
+    - name: Build and verify unsigned preview
+      if: runner.os != 'Linux'
+      shell: bash
+      working-directory: desktop
+      run: bun run make
+
+    - name: Build and verify unsigned preview under Xvfb
+      if: runner.os == 'Linux'
+      shell: bash
+      working-directory: desktop
+      run: |
+        bun run build
+        bun scripts/run-electron.mjs make
+        sudo chown root:root out/LeapView-linux-*/chrome-sandbox
+        sudo chmod 4755 out/LeapView-linux-*/chrome-sandbox
+        xvfb-run --auto-servernum node scripts/verify-package.mjs
+        node scripts/verify-installer.mjs
+
+    - name: Prove the packaged macOS preview against the hostile instance
+      if: runner.os == 'macOS'
+      shell: bash
+      run: |
+        executable="$(find desktop/out -path '*/LeapView.app/Contents/MacOS/LeapView' -print -quit)"
+        LEAPVIEW_PACKAGED_APP="$executable" go test -count=1 \
+          -run TestPackagedLeapViewPreservesRemoteContentBoundary \
+          -v ./internal/app/testing/maliciousinstance
+
+    - name: Prove the packaged Linux preview against the hostile instance
+      if: runner.os == 'Linux'
+      shell: bash
+      env:
+        LEAPVIEW_PACKAGED_APP: ${{ github.workspace }}/desktop/out/LeapView-linux-x64/LeapView
+      run: xvfb-run --auto-servernum go test -count=1 -run TestPackagedLeapViewPreservesRemoteContentBoundary -v ./internal/app/testing/maliciousinstance
+
+    - name: Prove the packaged Windows preview against the hostile instance
+      if: runner.os == 'Windows'
+      shell: pwsh
+      run: |
+        $env:LEAPVIEW_PACKAGED_APP = Get-ChildItem -Path desktop/out/LeapView-win32-* -Filter LeapView.exe |
+          Select-Object -First 1 -ExpandProperty FullName
+        go test -count=1 -run TestPackagedLeapViewPreservesRemoteContentBoundary -v ./internal/app/testing/maliciousinstance
+
+    - name: Generate and verify unsigned release evidence
+      shell: bash
+      working-directory: desktop
+      env:
+        GITHUB_SHA: ${{ inputs.source-sha }}
+      run: bun run evidence
+
+    - name: Qualify Linux installation lifecycle
+      if: runner.os == 'Linux'
+      shell: bash
+      working-directory: desktop
+      run: |
+        installer="$(find out/make -type f -name '*.deb' -print -quit)"
+        scripts/qualify-installer-linux.sh "$installer"
+
+    - name: Qualify macOS installation lifecycle
+      if: runner.os == 'macOS'
+      shell: bash
+      working-directory: desktop
+      run: |
+        installer="$(find out/make -type f -name '*.dmg' -print -quit)"
+        scripts/qualify-consumer-macos.sh "$installer"
+
+    - name: Qualify Windows installation lifecycle
+      if: runner.os == 'Windows'
+      shell: pwsh
+      working-directory: desktop
+      run: |
+        $installer = Get-ChildItem -Path out/make -Recurse -Filter *.exe |
+          Select-Object -First 1 -ExpandProperty FullName
+        ./scripts/qualify-consumer-windows.ps1 $installer
+
+    - name: Upload verifiable unsigned candidate
+      uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+      with:
+        name: ${{ inputs.artifact-name }}
+        retention-days: 7
+        if-no-files-found: error
+        path: |
+          desktop/out/make/**/*.zip
+          desktop/out/make/**/*.deb
+          desktop/out/make/**/*.dmg
+          desktop/out/make/**/*.exe
+          desktop/out/make/**/*.nupkg
+          desktop/out/make/**/RELEASES
+          desktop/out/evidence/**
+          desktop/release-policy.json
