+++ .github/workflows/desktop-preview-release.yml
@@ -11,7 +11,6 @@ on:
       release_tag:
         description: Immutable tag such as desktop-v0.1.0-alpha.1
         required: true
-        default: desktop-v0.1.0-alpha.1
         type: string
       confirm_unsigned_preview:
         description: Publish a public GitHub prerelease whose installers are unsigned
@@ -65,6 +64,16 @@ jobs:
           base_version="$(node -p 'require("./desktop/package.json").version')"
           version="$(node desktop/scripts/preview-release.mjs version \
             --tag "$RELEASE_TAG" --base-version "$base_version")"
+          manifest_tag="$(node -p 'require("./docs/desktop-release.json").release.tag')"
+          manifest_version="$(node -p 'require("./docs/desktop-release.json").release.version')"
+          test "$RELEASE_TAG" = "$manifest_tag" || {
+            printf 'Release tag %s does not match desktop manifest tag %s.\n' "$RELEASE_TAG" "$manifest_tag" >&2
+            exit 1
+          }
+          test "$version" = "$manifest_version" || {
+            printf 'Release version %s does not match desktop manifest version %s.\n' "$version" "$manifest_version" >&2
+            exit 1
+          }
           if git ls-remote --exit-code --tags origin "refs/tags/$RELEASE_TAG" >/dev/null 2>&1; then
             printf 'Tag %s already exists and cannot be replaced.\n' "$RELEASE_TAG" >&2
             exit 1
@@ -113,94 +122,11 @@ jobs:
         with:
           ref: ${{ needs.identity.outputs.source_sha }}
 
-      - name: Set up Bun
-        uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2
+      - name: Build, prove, qualify, and upload the exact preview candidate
+        uses: ./.github/actions/desktop-preview-candidate
         with:
-          bun-version: 1.3.7
-
-      - name: Set up Go for native Windows policy verification
-        if: runner.os == 'Windows'
-        uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16 # v6
-        with:
-          go-version-file: go.mod
-          cache: true
-
-      - name: Test native Windows policy probe
-        if: runner.os == 'Windows'
-        run: go test ./desktop/native/windowspolicy
-
-      - name: Install Debian package tooling
-        if: runner.os == 'Linux'
-        run: sudo apt-get update && sudo apt-get install --yes fakeroot
-
-      - name: Install pinned desktop dependencies
-        working-directory: desktop
-        run: bun install --frozen-lockfile
-
-      - name: Test preview and desktop security contracts
-        working-directory: desktop
-        run: bun run test
-
-      - name: Build and verify unsigned preview
-        if: runner.os != 'Linux'
-        working-directory: desktop
-        run: bun run make
-
-      - name: Build and verify unsigned preview under Xvfb
-        if: runner.os == 'Linux'
-        working-directory: desktop
-        run: |
-          bun run build
-          bun scripts/run-electron.mjs make
-          sudo chown root:root out/LeapView-linux-*/chrome-sandbox
-          sudo chmod 4755 out/LeapView-linux-*/chrome-sandbox
-          xvfb-run --auto-servernum node scripts/verify-package.mjs
-          node scripts/verify-installer.mjs
-
-      - name: Generate verifiable unsigned evidence
-        working-directory: desktop
-        env:
-          GITHUB_SHA: ${{ needs.identity.outputs.source_sha }}
-        run: bun run evidence
-
-      - name: Qualify Linux installation lifecycle
-        if: runner.os == 'Linux'
-        working-directory: desktop
-        run: |
-          installer="$(find out/make -type f -name '*.deb' -print -quit)"
-          scripts/qualify-installer-linux.sh "$installer"
-
-      - name: Qualify macOS installation lifecycle
-        if: runner.os == 'macOS'
-        working-directory: desktop
-        run: |
-          installer="$(find out/make -type f -name '*.dmg' -print -quit)"
-          scripts/qualify-consumer-macos.sh "$installer"
-
-      - name: Qualify Windows installation lifecycle
-        if: runner.os == 'Windows'
-        shell: pwsh
-        working-directory: desktop
-        run: |
-          $installer = Get-ChildItem -Path out/make -Recurse -Filter *.exe |
-            Select-Object -First 1 -ExpandProperty FullName
-          ./scripts/qualify-consumer-windows.ps1 $installer
-
-      - name: Upload immutable preview candidate
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: leapview-desktop-${{ matrix.artifact }}
-          retention-days: 7
-          if-no-files-found: error
-          path: |
-            desktop/out/make/**/*.zip
-            desktop/out/make/**/*.deb
-            desktop/out/make/**/*.dmg
-            desktop/out/make/**/*.exe
-            desktop/out/make/**/*.nupkg
-            desktop/out/make/**/RELEASES
-            desktop/out/evidence/**
-            desktop/release-policy.json
+          artifact-name: leapview-desktop-${{ matrix.artifact }}
+          source-sha: ${{ needs.identity.outputs.source_sha }}
 
   publish:
     name: Publish unsigned GitHub prerelease
