+++ .github/workflows/electron-security-proof.yml
@@ -2,6 +2,12 @@ name: Electron security proof
 
 on:
   pull_request:
+    paths:
+      - ".github/actions/desktop-preview-candidate/**"
+      - ".github/workflows/electron-security-proof.yml"
+      - "desktop/**"
+      - "internal/app/testing/maliciousinstance/**"
+      - "Taskfile.yml"
   push:
     branches:
       - main
@@ -32,6 +38,9 @@ jobs:
     name: Hardened package (${{ matrix.name }})
     runs-on: ${{ matrix.os }}
     timeout-minutes: 35
+    if: ${{ github.event_name != 'pull_request' || matrix.artifact == 'linux-x64' }}
+    env:
+      LEAPVIEW_DESKTOP_DISTRIBUTION: preview
     strategy:
       fail-fast: false
       matrix:
@@ -57,92 +66,11 @@ jobs:
       - name: Checkout
         uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
 
-      - name: Set up Bun
-        uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2
+      - name: Build, prove, qualify, and upload the exact preview candidate
+        uses: ./.github/actions/desktop-preview-candidate
         with:
-          bun-version: 1.3.7
-
-      - name: Set up Go for native Windows policy verification
-        if: runner.os == 'Windows'
-        uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16 # v6
-        with:
-          go-version-file: go.mod
-          cache: true
-
-      - name: Test native Windows policy probe
-        if: runner.os == 'Windows'
-        run: go test ./desktop/native/windowspolicy
-
-      - name: Install Debian package tooling
-        if: runner.os == 'Linux'
-        run: sudo apt-get update && sudo apt-get install --yes fakeroot
-
-      - name: Install pinned desktop dependencies
-        working-directory: desktop
-        run: bun install --frozen-lockfile
-
-      - name: Test the trusted desktop and updater contracts
-        working-directory: desktop
-        run: bun run test
-
-      - name: Build, harden, inspect, and launch the package
-        if: runner.os != 'Linux'
-        working-directory: desktop
-        run: bun run make
-
-      - name: Build, harden, inspect, and launch the package under Xvfb
-        if: runner.os == 'Linux'
-        working-directory: desktop
-        run: |
-          bun run build
-          bun scripts/run-electron.mjs make
-          sudo chown root:root out/LeapView-linux-*/chrome-sandbox
-          sudo chmod 4755 out/LeapView-linux-*/chrome-sandbox
-          xvfb-run --auto-servernum node scripts/verify-package.mjs
-          node scripts/verify-installer.mjs
-
-      - name: Generate and verify release evidence
-        working-directory: desktop
-        run: bun run evidence
-
-      - name: Qualify Linux install, upgrade, protocol, and uninstall
-        if: runner.os == 'Linux'
-        working-directory: desktop
-        run: |
-          installer="$(find out/make -type f -name '*.deb' -print -quit)"
-          scripts/qualify-installer-linux.sh "$installer"
-
-      - name: Qualify macOS consumer install, protocol, and removal
-        if: runner.os == 'macOS'
-        working-directory: desktop
-        run: |
-          installer="$(find out/make -type f -name '*.dmg' -print -quit)"
-          scripts/qualify-consumer-macos.sh "$installer"
-
-      - name: Qualify Windows per-user install, protocol, and uninstall
-        if: runner.os == 'Windows'
-        shell: pwsh
-        working-directory: desktop
-        run: |
-          $installer = Get-ChildItem -Path out/make -Recurse -Filter *.exe |
-            Select-Object -First 1 -ExpandProperty FullName
-          ./scripts/qualify-consumer-windows.ps1 $installer
-
-      - name: Upload verifiable unsigned candidate
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: leapview-desktop-${{ matrix.artifact }}
-          retention-days: 7
-          if-no-files-found: error
-          path: |
-            desktop/out/make/**/*.zip
-            desktop/out/make/**/*.deb
-            desktop/out/make/**/*.dmg
-            desktop/out/make/**/*.exe
-            desktop/out/make/**/*.nupkg
-            desktop/out/make/**/RELEASES
-            desktop/out/evidence/**
-            desktop/release-policy.json
+          artifact-name: leapview-desktop-${{ matrix.artifact }}
+          source-sha: ${{ github.sha }}
 
   attest:
     name: Attest main candidate (${{ matrix.name }})
@@ -221,7 +149,8 @@ jobs:
           sbom-path: candidate/out/evidence/*.spdx.json
 
   macos:
-    name: Candidate proof (macOS ${{ matrix.architecture }})
+    name: Policy integration (macOS ${{ matrix.architecture }})
+    if: ${{ github.event_name != 'pull_request' }}
     strategy:
       fail-fast: false
       matrix:
@@ -259,11 +188,12 @@ jobs:
           LEAPVIEW_ELECTRON_BINARY: ${{ github.workspace }}/internal/app/testing/maliciousinstance/electron/node_modules/electron/dist/Electron.app/Contents/MacOS/Electron
         run: |
           go test -count=1 \
-            -run TestElectronCandidatePreservesBrowserEquivalentAuthority \
+            -run TestElectronPolicyIntegrationPreservesBrowserEquivalentAuthority \
             -v ./internal/app/testing/maliciousinstance
 
   windows:
-    name: Candidate proof (Windows x64)
+    name: Policy integration (Windows x64)
+    if: ${{ github.event_name != 'pull_request' }}
     runs-on: windows-2025
     timeout-minutes: 15
 
@@ -292,10 +222,10 @@ jobs:
         shell: pwsh
         run: |
           $env:LEAPVIEW_ELECTRON_BINARY = Join-Path $env:GITHUB_WORKSPACE "internal/app/testing/maliciousinstance/electron/node_modules/electron/dist/electron.exe"
-          go test -count=1 -run TestElectronCandidatePreservesBrowserEquivalentAuthority -v ./internal/app/testing/maliciousinstance
+          go test -count=1 -run TestElectronPolicyIntegrationPreservesBrowserEquivalentAuthority -v ./internal/app/testing/maliciousinstance
 
   linux:
-    name: Candidate proof (Linux x64)
+    name: Policy integration (Linux x64)
     runs-on: ubuntu-24.04
     timeout-minutes: 15
