+++ desktop/src/distribution.ts
@@ -2,7 +2,9 @@ import { readFileSync } from "node:fs";
 import { join } from "node:path";
 
 export const PREVIEW_DISTRIBUTION_MARKER = "preview-distribution.json";
+export const STABLE_DISTRIBUTION_MARKER = "stable-distribution.json";
 const previewMarker = '{"schemaVersion":1,"channel":"preview","updates":false}\n';
+const stableMarker = '{"schemaVersion":1,"channel":"stable","updates":true}\n';
 
 export type DesktopDistribution =
   | "development"
@@ -24,16 +26,34 @@ export function resolveDesktopDistribution(
   }
   const readFile =
     options.readFile ?? ((path: string) => readFileSync(path, "utf8"));
+  const preview = readMarker(
+    readFile,
+    join(options.resourcesPath, PREVIEW_DISTRIBUTION_MARKER),
+  );
+  const stable = readMarker(
+    readFile,
+    join(options.resourcesPath, STABLE_DISTRIBUTION_MARKER),
+  );
+  if (preview === previewMarker && stable === undefined) {
+    return "preview";
+  }
+  if (stable === stableMarker && preview === undefined) {
+    return "stable";
+  }
+  return "invalid";
+}
+
+function readMarker(
+  readFile: (path: string) => string,
+  path: string,
+): string | undefined | null {
   try {
-    return readFile(join(options.resourcesPath, PREVIEW_DISTRIBUTION_MARKER)) ===
-        previewMarker
-      ? "preview"
-      : "invalid";
+    return readFile(path);
   } catch (error) {
     return error instanceof Error &&
         "code" in error &&
         error.code === "ENOENT"
-      ? "stable"
-      : "invalid";
+      ? undefined
+      : null;
   }
 }
