+++ desktop/src/security/remote-policy.mjs
@@ -79,18 +79,6 @@ export function configureRemoteSession(
   audit = () => {},
   capabilities = {},
 ) {
-  remoteSession.webRequest.onBeforeRequest(
-    {
-      urls: [
-        "http://wails.localhost/*",
-        "https://wails.localhost/*",
-      ],
-    },
-    (_details, callback) => {
-      audit({ kind: "native-transport", allowed: false });
-      callback({ cancel: true });
-    },
-  );
   remoteSession.setPermissionCheckHandler((_contents, permission) => {
     audit({ kind: "permission-check", permission, allowed: false });
     return false;
