+++ desktop/stable-distribution.json
@@ -1,0 +1,1 @@
+{"schemaVersion":1,"channel":"stable","updates":true}
