+++ docs/articles/desktop/install.md
@@ -2,24 +2,26 @@
 
 End users install LeapView Desktop from the official [download page](/download). They do not clone the repository, run build commands, or need source code and development tools.
 
-> **Availability:** Production installers are not published yet. The download page will not expose unsigned CI candidates while signing and final qualification are incomplete.
+> **Availability:** The current download is an unsigned early preview published through the verified LeapView GitHub prerelease. macOS and Windows publisher warnings are expected. Signed production installers remain a later release milestone.
 
 ## Choose the right installer
 
 | System | Supported version | Architecture | Installer |
 | --- | --- | --- | --- |
-| macOS | macOS 13 Ventura or newer | Intel or Apple silicon | signed and notarized DMG |
-| Windows | Windows 10 or newer | x64 | signed per-user Squirrel Setup EXE |
-| Ubuntu | Ubuntu 22.04 LTS or newer | x64 | DEB from the signed LeapView APT repository |
+| macOS | macOS 13 Ventura or newer | Intel or Apple silicon | preview DMG |
+| Windows | Windows 10 or newer | x64 | preview per-user Squirrel Setup EXE |
+| Ubuntu | Ubuntu 22.04 LTS or newer | x64 | preview DEB |
 
-The download page publishes the version, architecture, minimum operating-system version, SHA-256 checksum, code-signature identity, SBOM, provenance, and release notes beside each production artifact. Unsupported system and architecture combinations do not receive a generic or best-effort download.
+The download page and its linked GitHub release identify the version, architecture, minimum operating-system version, immutable installer names, SHA-256 checksums, SBOMs, provenance, and signing state. Unsupported system and architecture combinations do not receive a generic or best-effort download.
+
+The production formats remain a signed and notarized DMG, a signed per-user Squirrel Setup EXE, and the signed LeapView APT repository. Those stable artifacts appear on the same download page after qualification.
 
 ## macOS
 
 1. Download the DMG matching the Mac architecture.
 2. Open it and drag LeapView to Applications.
 3. Open LeapView from Applications.
-4. Confirm that macOS identifies the verified LeapView publisher. Do not bypass a Gatekeeper warning.
+4. For the unsigned preview, use **Control-click → Open** or approve the application in **Privacy & Security** only after verifying the release evidence. A future stable release must identify the verified LeapView publisher without this exception.
 
 The DMG is the human installation format. A matching ZIP exists only for the built-in Squirrel.Mac updater and is not the normal installation choice.
 
@@ -29,7 +31,7 @@ To uninstall, quit LeapView and move it from Applications to the Bin. Removing t
 
 1. Download the x64 Setup EXE.
 2. Run the installer as the intended user.
-3. Confirm that Windows identifies the verified LeapView publisher. Do not bypass a SmartScreen signature failure.
+3. The unsigned preview may show Microsoft Defender SmartScreen. Continue only after checking that the installer came from the linked LeapView prerelease and its SHA-256 matches `SHA256SUMS`.
 4. Launch LeapView from the Start menu or desktop shortcut.
 
 The Squirrel installer is per-user and does not require an administrator for routine installation or updates. Its NUPKG and `RELEASES` companions are updater inputs, not separate installers.
@@ -38,7 +40,7 @@ Use **Settings > Apps > Installed apps > LeapView > Uninstall** to remove it. Th
 
 ## Ubuntu
 
-Install LeapView from the signed LeapView APT repository using the commands shown on the download page. APT verifies repository metadata, selects the x64 DEB, installs desktop and URL-handler registration, and owns later upgrades.
+Download the preview DEB from the linked LeapView prerelease and verify it against `SHA256SUMS`, then install it with the operating-system package manager. The stable Linux release will move to the signed LeapView APT repository.
 
 Use the operating-system package manager to uninstall:
