+++ docs/desktop-release.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
-  "status": "preparing",
+  "status": "published",
   "product": {
     "name": "LeapView",
     "applicationId": "dev.leapview.desktop"
   },
   "channel": {
-    "name": "stable",
-    "updateOrigin": "https://releases.leapview.dev",
+    "name": "preview",
+    "updateOrigin": "",
     "pathVersion": "v1"
   },
   "support": [
@@ -34,5 +34,57 @@
       "minimumVersion": "Windows 10"
     }
   ],
-  "release": null
+  "release": {
+    "version": "0.1.0-alpha.1",
+    "tag": "desktop-v0.1.0-alpha.1",
+    "notesUrl": "https://github.com/flidai/leapview/releases/tag/desktop-v0.1.0-alpha.1",
+    "evidenceUrl": "https://github.com/flidai/leapview/releases/tag/desktop-v0.1.0-alpha.1",
+    "signingStatus": "unsigned",
+    "artifacts": [
+      {
+        "platform": "darwin",
+        "architecture": "arm64",
+        "format": "dmg",
+        "fileName": "LeapView-Desktop-0.1.0-alpha.1-macos-arm64.dmg",
+        "downloadUrl": "https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-macos-arm64.dmg",
+        "signature": {
+          "type": "unsigned",
+          "identity": "Unsigned early preview"
+        }
+      },
+      {
+        "platform": "darwin",
+        "architecture": "x64",
+        "format": "dmg",
+        "fileName": "LeapView-Desktop-0.1.0-alpha.1-macos-x64.dmg",
+        "downloadUrl": "https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-macos-x64.dmg",
+        "signature": {
+          "type": "unsigned",
+          "identity": "Unsigned early preview"
+        }
+      },
+      {
+        "platform": "win32",
+        "architecture": "x64",
+        "format": "exe",
+        "fileName": "LeapView-Desktop-0.1.0-alpha.1-windows-x64.exe",
+        "downloadUrl": "https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-windows-x64.exe",
+        "signature": {
+          "type": "unsigned",
+          "identity": "Unsigned early preview"
+        }
+      },
+      {
+        "platform": "linux",
+        "architecture": "x64",
+        "format": "deb",
+        "fileName": "LeapView-Desktop-0.1.0-alpha.1-linux-x64.deb",
+        "downloadUrl": "https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-linux-x64.deb",
+        "signature": {
+          "type": "unsigned",
+          "identity": "Unsigned early preview"
+        }
+      }
+    ]
+  }
 }
