+++ internal/access/sqlite/queries/access.sql
@@ -298,6 +298,58 @@ UPDATE sessions
 SET revoked_at = COALESCE(revoked_at, CURRENT_TIMESTAMP)
 WHERE principal_id = ? AND revoked_at IS NULL;
 
+-- name: CleanDesktopAuthorizationCodes :exec
+DELETE FROM desktop_authorization_codes
+WHERE expires_at <= sqlc.arg(now) OR consumed_at IS NOT NULL;
+
+-- name: CreateDesktopAuthorizationCode :exec
+INSERT INTO desktop_authorization_codes (
+  code_hash, principal_id, client_id, instance_id, profile_id,
+  redirect_uri, code_challenge, return_path, expires_at, consumed_at, created_at
+) VALUES (
+  sqlc.arg(code_hash), sqlc.arg(principal_id), sqlc.arg(client_id),
+  sqlc.arg(instance_id), sqlc.arg(profile_id), sqlc.arg(redirect_uri),
+  sqlc.arg(code_challenge), sqlc.arg(return_path), sqlc.arg(expires_at),
+  NULL, sqlc.arg(created_at)
+);
+
+-- name: GetDesktopAuthorizationCode :one
+SELECT code_hash, principal_id, client_id, instance_id, profile_id,
+       redirect_uri, code_challenge, return_path, expires_at, consumed_at,
+       created_at
+FROM desktop_authorization_codes
+WHERE code_hash = sqlc.arg(code_hash);
+
+-- name: ConsumeDesktopAuthorizationCode :execrows
+UPDATE desktop_authorization_codes
+SET consumed_at = sqlc.arg(consumed_at)
+WHERE code_hash = sqlc.arg(code_hash) AND consumed_at IS NULL;
+
+-- name: CreateDesktopSessionBinding :exec
+INSERT INTO desktop_sessions (
+  session_id, instance_id, profile_id, client_id, absolute_expires_at, created_at
+) VALUES (
+  sqlc.arg(session_id), sqlc.arg(instance_id), sqlc.arg(profile_id),
+  sqlc.arg(client_id), sqlc.arg(absolute_expires_at), sqlc.arg(created_at)
+);
+
+-- name: GetDesktopSessionByTokenFingerprint :one
+SELECT s.id, s.principal_id, s.token_verifier, s.expires_at,
+       ds.instance_id, ds.profile_id, ds.client_id,
+       ds.absolute_expires_at, ds.created_at
+FROM sessions s
+JOIN desktop_sessions ds ON ds.session_id = s.id
+WHERE s.token_fingerprint = sqlc.arg(token_fingerprint)
+  AND s.revoked_at IS NULL
+  AND datetime(s.expires_at) > CURRENT_TIMESTAMP
+  AND datetime(ds.absolute_expires_at) > CURRENT_TIMESTAMP
+  AND datetime(s.last_seen_at) > datetime(sqlc.arg(idle_cutoff));
+
+-- name: RevokeActiveSessionByID :execrows
+UPDATE sessions
+SET revoked_at = COALESCE(revoked_at, CURRENT_TIMESTAMP)
+WHERE id = sqlc.arg(id) AND revoked_at IS NULL;
+
 -- name: CreateAPIToken :exec
 INSERT INTO api_tokens (id, principal_id, workspace_id, name, token_fingerprint, token_verifier, privileges_json, expires_at)
 VALUES (?, ?, ?, ?, ?, ?, ?, ?);
