+++ internal/app/config/spec/spec.go
@@ -191,7 +191,10 @@ var settings = []Setting{
 	{Name: "LEAPVIEW_PRODUCTION", Field: "Production", Type: TypeBool, Category: "server", Scope: "serve,admin", Description: "Enable production serving and validation behavior.", Default: "false", Runtime: true, Lifecycle: "supported", EnvExample: "1"},
 	{Name: "LEAPVIEW_PUBLIC_URL", Field: "PublicURL", Type: TypeString, Category: "server", Scope: "serve", Description: "Canonical externally visible LeapView origin used for MCP resource identity and OAuth discovery.", Example: "https://leapview.example.com", Runtime: true, Lifecycle: "supported", EnvExample: "https://leapview.example.com"},
 	{Name: "LEAPVIEW_PUBLIC_RELEASE_MANIFEST", Type: TypeString, Default: "docs/public-release.json", Category: "release", Scope: "public site smoke", Description: "Release manifest compared with the public site's deployed release identity.", Lifecycle: "tooling"},
+	{Name: "LEAPVIEW_DESKTOP_DISTRIBUTION", Type: TypeString, Category: "release", Scope: "desktop packaging", Description: "Required desktop package identity; accepted values are preview and stable.", Example: "preview", Lifecycle: "tooling"},
+	{Name: "LEAPVIEW_DESKTOP_PACKAGED_PROOF_ORIGIN", Type: TypeString, Category: "release", Scope: "packaged desktop security proof", Description: "Exact loopback hostile-instance origin injected only into an isolated packaged-app security proof.", Example: "http://127.0.0.1:18080", Lifecycle: "internal"},
 	{Name: "LEAPVIEW_DESKTOP_RELEASE_MANIFEST", Type: TypeString, Default: "docs/desktop-release.json", Category: "release", Scope: "public site smoke", Description: "Desktop release manifest compared with the public download page and deployed desktop release identity.", Lifecycle: "tooling"},
+	{Name: "LEAPVIEW_PACKAGED_APP", Type: TypeString, Category: "release", Scope: "packaged desktop security proof", Description: "Path to the exact packaged LeapView executable qualified by the hostile-instance proof.", Example: "desktop/out/LeapView-linux-x64/LeapView", Lifecycle: "internal"},
 	{Name: "LEAPVIEW_PUBLIC_SITE_ALIASES", Type: TypeString, Default: "http://leapview.dev,https://www.leapview.dev", Category: "release", Scope: "public site smoke", Description: "Comma-separated public aliases that must redirect to the canonical site origin.", Lifecycle: "tooling"},
 	{Name: "LEAPVIEW_PUBLIC_SITE_URL", Type: TypeString, Default: "https://leapview.dev", Category: "release", Scope: "public site smoke", Description: "Canonical public site origin verified by the release adoption smoke test.", Lifecycle: "tooling"},
 	{Name: "LEAPVIEW_SCIM_BEARER_TOKEN", Field: "SCIMBearerToken", Type: TypeString, Category: "authentication", Scope: "serve", Description: "Bearer token enabling SCIM provisioning; production requires at least 32 characters when set.", Example: SecretPlaceholder, Secret: true, Runtime: true, Lifecycle: "supported", Commented: true},
