+++ internal/app/testing/maliciousinstance/README.md
@@ -50,7 +50,7 @@ Run the macOS integration proof from the repository root:
 
 ```sh
 LEAPVIEW_ELECTRON_BINARY="$PWD/internal/app/testing/maliciousinstance/electron/node_modules/electron/dist/Electron.app/Contents/MacOS/Electron" \
-  go test -count=1 -run TestElectronCandidatePreservesBrowserEquivalentAuthority -v \
+  go test -count=1 -run TestElectronPolicyIntegrationPreservesBrowserEquivalentAuthority -v \
   ./internal/app/testing/maliciousinstance
 ```
 
@@ -68,7 +68,7 @@ Do not substitute Electron's `--no-sandbox` flag. It invalidates the proof.
 
 The Go integration test skips unless `LEAPVIEW_ELECTRON_BINARY` points to an
 explicit Electron binary, keeping ordinary unit-test runs fast and deterministic.
-The `Electron security proof` workflow runs the same complete 23-attack
+The `Electron security proof` workflow runs the same complete 20-invariant
 manifest on macOS Intel, Windows x64, and sandboxed Linux x64. Proof output
 contains only bounded enum-like observations and framework versions; it must
 not contain credentials or tenant data.
