+++ internal/app/testing/maliciousinstance/electron/policy.test.mjs
@@ -87,22 +87,7 @@ test("remote session denies permissions, devices, and downloads", () => {
   session.emit("will-download", event, item, {});
   assert.equal(event.defaultPrevented, true);
   assert.equal(item.cancelCalled, true);
-  assert.deepEqual(session.beforeRequestFilter, {
-    urls: [
-      "http://wails.localhost/*",
-      "https://wails.localhost/*",
-    ],
-  });
-  let transportDecision;
-  session.beforeRequestHandler(
-    { url: "http://wails.localhost/wails/runtime" },
-    (decision) => {
-      transportDecision = decision;
-    },
-  );
-  assert.deepEqual(transportDecision, { cancel: true });
   assert.ok(decisions.some((decision) => decision.kind === "download" && decision.allowed === false));
-  assert.ok(decisions.some((decision) => decision.kind === "native-transport" && decision.allowed === false));
 });
 
 test("reviewed CSV exports require an exact-profile origin, user gesture, and save dialog", () => {
