+++ site/site.dom.test.ts
@@ -488,18 +488,19 @@ test('site brand pairs the LeapView wordmark with the Lucide Aperture ring mark'
   }
 })
 
-test('desktop download page remains accessible before publication', async () => {
+test('desktop download page presents the same manifest-backed early preview', async () => {
   const page = await browser.newPage({ viewport: { width: 1280, height: 900 } })
   try {
     await page.goto(`${baseURL}/download`)
 
     expect(await page.getByRole('heading', { level: 1, name: 'LeapView on your desktop.' }).isVisible()).toBe(true)
     expect(await page.locator('.site-download-hero > .site-eyebrow').count()).toBe(0)
-    expect(await page.getByText('Production downloads are not published yet.').isVisible()).toBe(true)
+    expect(await page.getByText('Early preview', { exact: true }).isVisible()).toBe(true)
+    expect(await page.getByText('These installers are not code-signed.', { exact: false }).isVisible()).toBe(true)
     expect(await page.getByRole('link', { name: 'Read the install guide' }).getAttribute('href')).toBe('/docs/desktop/install')
     expect(await page.getByRole('link', { name: 'Review desktop security' }).getAttribute('href')).toBe('/docs/desktop/security')
     expect(await page.locator('.site-download-platform').count()).toBe(3)
-    expect(await page.locator('a[download]').count()).toBe(0)
+    expect(await page.locator('.site-download-artifact .site-button-primary').count()).toBe(4)
   } finally {
     await page.close()
   }
@@ -508,6 +509,24 @@ test('desktop download page remains accessible before publication', async () => 
 test('homepage offers the attested unsigned desktop preview for all platforms', async () => {
   const page = await browser.newPage({ viewport: { width: 1440, height: 1000 } })
   try {
+    const releaseResponse = await page.request.get(`${baseURL}/desktop-release.json`)
+    const releaseManifest = await releaseResponse.json() as {
+      release: {
+        evidenceUrl: string
+        artifacts: Array<{
+          platform: string
+          architecture: string
+          downloadUrl: string
+        }>
+      }
+    }
+    const artifactURL = (platform: string, architecture: string) => {
+      const artifact = releaseManifest.release.artifacts.find(
+        (candidate) => candidate.platform === platform && candidate.architecture === architecture,
+      )
+      expect(artifact).toBeDefined()
+      return artifact!.downloadUrl
+    }
     await page.goto(baseURL)
     const section = page.locator('.site-desktop-section')
     expect(await section.getByRole('heading', { level: 2, name: 'Take LeapView to your desktop.' }).isVisible()).toBe(true)
@@ -595,9 +614,9 @@ test('homepage offers the attested unsigned desktop preview for all platforms', 
     const cards = section.locator('.site-desktop-platform')
     expect(await cards.count()).toBe(3)
     for (const [platform, label, href] of [
-      ['macos', 'Download for macOS', 'https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-macos-arm64.dmg'],
-      ['windows', 'Download for Windows', 'https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-windows-x64.exe'],
-      ['linux', 'Download for Linux', 'https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-linux-x64.deb'],
+      ['macos', 'Download for macOS', artifactURL('darwin', 'arm64')],
+      ['windows', 'Download for Windows', artifactURL('win32', 'x64')],
+      ['linux', 'Download for Linux', artifactURL('linux', 'x64')],
     ] as const) {
       const card = section.locator(`.site-desktop-platform[data-desktop-platform="${platform}"]`)
       expect(await card.count()).toBe(1)
@@ -610,9 +629,9 @@ test('homepage offers the attested unsigned desktop preview for all platforms', 
     }
     expect(
       await section.getByRole('link', { name: 'Download for Intel Mac' }).getAttribute('href'),
-    ).toBe('https://github.com/flidai/leapview/releases/download/desktop-v0.1.0-alpha.1/LeapView-Desktop-0.1.0-alpha.1-macos-x64.dmg')
+    ).toBe(artifactURL('darwin', 'x64'))
     expect(await desktopSubtitle.getByRole('link', { name: 'Verify release evidence' }).getAttribute('href')).toBe(
-      'https://github.com/flidai/leapview/releases/tag/desktop-v0.1.0-alpha.1',
+      releaseManifest.release.evidenceUrl,
     )
     await page.setViewportSize({ width: 390, height: 844 })
     await page.reload()
